requirements.txt
discord.py
python-dotenv
flask
gunicorn
pytz
pandas
supabase